rem trainer.py and update auto_shape
