--- doc/code/training/trainer.rst
@@ -1,7 +1,0 @@
-trainer
-============
-
-.. currentmodule:: mrmustard.training.trainer
-
-.. automodapi:: mrmustard.training.trainer
-    :no-heading:
