+++ mrmustard/math/lattice/autoshape.py
@@ -149,9 +149,7 @@ def autoshape_numba(A, b, c, max_prob, max_shape, min_shape) -> int:  # pragma: 
             ) / SQRT[k + 2]
             norm += np.abs(buf3[(k + 1) % 2, 1])
             k += 1
-        shape[m] = k or 1
-        if shape[m] < min_shape:
-            shape[m] = min_shape
-        # if shape[m] > max_shape:
-        #     shape[m] = max_shape
+
+        shape[m] = k
+    shape = np.clip(shape, min_shape, max_shape)
     return shape
